make StepSlider an indepedent component
