+++ src/css/media.less
@@ -306,7 +306,7 @@
 		.head-desc {
 		  h1 {
 			.font-size(5.3);
-			line-height: 50px;
+			 line-height: 60px;
 		  }
 
 		  .details {
