+++ src/css/style.less
@@ -3150,10 +3150,15 @@ form[name='compose'] {
           }
 
           .icon {
-            .font-size(9.0);
+            .font-size(9);
             color: @primary;
             padding: 10px 0;
-            height: 120px;
+
+            span {
+              color: #F6F6F6;
+              font-weight: 400;
+              .font-size(11);
+            }
 
             img {
               vertical-align: top;
@@ -3444,7 +3449,7 @@ form[name='compose'] {
           padding: 0 10px;
 
           .icon {
-            .font-size(10);
+            .font-size(9);
           }
         }
       }
